src/core/multimodal/multimodal_processor.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
多模态处理模块
负责处理不同类型的输入（文本、图像、语音等）
"""

from abc import ABC, abstractmethod
from typing import Dict, List, Optional, Any, Union
from PIL import Image
import numpy as np


class MultimodalProcessor(ABC):
    """多模态处理器接口
    
    定义多模态处理的统一接口
    """
    
    @abstractmethod
    def process(self, input_data: Union[str, Image.Image, np.ndarray, bytes], input_type: str, task: str) -> Dict[str, Any]:
        """处理多模态输入
        
        Args:
            input_data: 输入数据
            input_type: 输入类型，如"text", "image", "audio"
            task: 任务类型，如"classification", "detection", "description", "generation"
            
        Returns:
            处理结果
        """
        pass
    
    @abstractmethod
    def get_supported_input_types(self) -> List[str]:
        """获取支持的输入类型
        
        Returns:
            支持的输入类型列表
        """
        pass
    
    @abstractmethod
    def get_supported_tasks(self) -> List[str]:
        """获取支持的任务类型
        
        Returns:
            支持的任务类型列表
        """
        pass


class BasicMultimodalProcessor(MultimodalProcessor):
    """基础多模态处理器
    
    实现基本的多模态处理功能
    """
    
    def __init__(self):
        """初始化多模态处理器"""
        # 支持的输入类型
        self.supported_input_types = ["text", "image", "audio"]
        # 支持的任务类型
        self.supported_tasks = ["classification", "detection", "description", "generation"]
    
    def process(self, input_data: Union[str, Image.Image, np.ndarray, bytes], input_type: str, task: str) -> Dict[str, Any]:
        """处理多模态输入
        
        Args:
            input_data: 输入数据
            input_type: 输入类型，如"text", "image", "audio"
            task: 任务类型，如"classification", "detection", "description", "generation"
            
        Returns:
            处理结果
        """
        # 检查输入类型是否支持
        if input_type not in self.supported_input_types:
            raise ValueError(f"Unsupported input type: {input_type}")
        
        # 检查任务类型是否支持
        if task not in self.supported_tasks:
            raise ValueError(f"Unsupported task type: {task}")
        
        # 根据输入类型和任务类型进行处理
        if input_type == "text":
            return self._process_text(input_data, task)
        elif input_type == "image":
            return self._process_image(input_data, task)
        elif input_type == "audio":
            return self._process_audio(input_data, task)
        else:
            raise ValueError(f"Unsupported input type: {input_type}")
    
    def _process_text(self, text: str, task: str) -> Dict[str, Any]:
        """处理文本输入
        
        Args:
            text: 文本输入
            task: 任务类型
            
        Returns:
            处理结果
        """
        if task == "classification":
            # 模拟文本分类
            return {
                "classification": "positive",
                "confidence": 0.85,
                "input_type": "text",
                "task": task
            }
        elif task == "generation":
            # 模拟文本生成
            return {
                "generated_text": f"Generated response to: {text}",
                "input_type": "text",
                "task": task
            }
        else:
            return {
                "message": f"Task {task} not supported for text input",
                "input_type": "text",
                "task": task
            }
    
    def _process_image(self, image: Union[Image.Image, np.ndarray], task: str) -> Dict[str, Any]:
        """处理图像输入
        
        Args:
            image: 图像输入
            task: 任务类型
            
        Returns:
            处理结果
        """
        if task == "classification":
            # 模拟图像分类
            return {
                "classification": "cat",
                "confidence": 0.95,
                "input_type": "image",
                "task": task
            }
        elif task == "detection":
            # 模拟目标检测
            return {
                "objects": [
                    {"class": "person", "confidence": 0.98, "bbox": [100, 100, 200, 300]},
                    {"class": "car", "confidence": 0.92, "bbox": [300, 200, 500, 350]}
                ],
                "input_type": "image",
                "task": task
            }
        elif task == "description":
            # 模拟图像描述
            return {
                "description": "A cat sitting on a couch",
                "input_type": "image",
                "task": task
            }
        else:
            return {
                "message": f"Task {task} not supported for image input",
                "input_type": "image",
                "task": task
            }
    
    def _process_audio(self, audio: bytes, task: str) -> Dict[str, Any]:
        """处理音频输入
        
        Args:
            audio: 音频输入
            task: 任务类型
            
        Returns:
            处理结果
        """
        if task == "classification":
            # 模拟音频分类
            return {
                "classification": "speech",
                "confidence": 0.90,
                "input_type": "audio",
                "task": task
            }
        elif task == "description":
            # 模拟音频描述
            return {
                "description": "A person speaking English",
                "input_type": "audio",
                "task": task
            }
        else:
            return {
                "message": f"Task {task} not supported for audio input",
                "input_type": "audio",
                "task": task
            }
    
    def get_supported_input_types(self) -> List[str]:
        """获取支持的输入类型
        
        Returns:
            支持的输入类型列表
        """
        return self.supported_input_types
    
    def get_supported_tasks(self) -> List[str]:
        """获取支持的任务类型
        
        Returns:
            支持的任务类型列表
        """
        return self.supported_tasks